Move Scenes to "Scenes" folder Move GameObjects to "GameObjects" folder
