+++ NonEuclidean/Engine.h
@@ -3,8 +3,8 @@
 #include "Camera.h"
 #include "Input.h"
 #include "Object.h"
-#include "Portal.h"
-#include "Player.h"
+#include "GameObjects/Portal.h"
+#include "GameObjects/Player.h"
 #include "Timer.h"
 #include "Scene.h"
 #include "Sky.h"
