+++ NonEuclidean/GameObjects/Portal.cpp
@@ -1,5 +1,5 @@
 #include "Portal.h"
-#include "Engine.h"
+#include "../Engine.h"
 #include <cassert>
 #include <iostream>
