+++ NonEuclidean/NonEuclidean.vcxproj
@@ -156,18 +156,18 @@
     <ClCompile Include="Engine.cpp" />
     <ClCompile Include="FrameBuffer.cpp" />
     <ClCompile Include="Input.cpp" />
-    <ClCompile Include="Level1.cpp" />
-    <ClCompile Include="Level2.cpp" />
-    <ClCompile Include="Level3.cpp" />
-    <ClCompile Include="Level4.cpp" />
-    <ClCompile Include="Level5.cpp" />
-    <ClCompile Include="Level6.cpp" />
+    <ClCompile Include="Scenes\Level1.cpp" />
+    <ClCompile Include="Scenes\Level2.cpp" />
+    <ClCompile Include="Scenes\Level3.cpp" />
+    <ClCompile Include="Scenes\Level4.cpp" />
+    <ClCompile Include="Scenes\Level5.cpp" />
+    <ClCompile Include="Scenes\Level6.cpp" />
     <ClCompile Include="Main.cpp" />
     <ClCompile Include="Mesh.cpp" />
     <ClCompile Include="Object.cpp" />
     <ClCompile Include="Physical.cpp" />
-    <ClCompile Include="Player.cpp" />
-    <ClCompile Include="Portal.cpp" />
+    <ClCompile Include="GameObjects\Player.cpp" />
+    <ClCompile Include="GameObjects\Portal.cpp" />
     <ClCompile Include="Resources.cpp" />
     <ClCompile Include="Shader.cpp" />
     <ClCompile Include="Texture.cpp" />
@@ -175,39 +175,39 @@
   <ItemGroup>
     <ClInclude Include="Camera.h" />
     <ClInclude Include="Collider.h" />
-    <ClInclude Include="Floorplan.h" />
-    <ClInclude Include="Ground.h" />
-    <ClInclude Include="House.h" />
+    <ClInclude Include="GameObjects\Floorplan.h" />
+    <ClInclude Include="GameObjects\Ground.h" />
+    <ClInclude Include="GameObjects\House.h" />
     <ClInclude Include="Engine.h" />
     <ClInclude Include="FrameBuffer.h" />
     <ClInclude Include="GameHeader.h" />
     <ClInclude Include="Input.h" />
-    <ClInclude Include="Level1.h" />
-    <ClInclude Include="Level2.h" />
-    <ClInclude Include="Level3.h" />
-    <ClInclude Include="Level4.h" />
-    <ClInclude Include="Level5.h" />
-    <ClInclude Include="Level6.h" />
+    <ClInclude Include="Scenes\Level1.h" />
+    <ClInclude Include="Scenes\Level2.h" />
+    <ClInclude Include="Scenes\Level3.h" />
+    <ClInclude Include="Scenes\Level4.h" />
+    <ClInclude Include="Scenes\Level5.h" />
+    <ClInclude Include="Scenes\Level6.h" />
     <ClInclude Include="Mesh.h" />
     <ClInclude Include="Meshes\Meshes.h" />
     <ClInclude Include="Object.h" />
     <ClInclude Include="Physical.h" />
-    <ClInclude Include="Pillar.h" />
-    <ClInclude Include="PillarRoom.h" />
-    <ClInclude Include="Player.h" />
-    <ClInclude Include="Portal.h" />
+    <ClInclude Include="GameObjects\Pillar.h" />
+    <ClInclude Include="GameObjects\PillarRoom.h" />
+    <ClInclude Include="GameObjects\Player.h" />
+    <ClInclude Include="GameObjects\Portal.h" />
     <ClInclude Include="Resources.h" />
     <ClInclude Include="Scene.h" />
     <ClInclude Include="Shader.h" />
     <ClInclude Include="Shaders\Shaders.h" />
     <ClInclude Include="Sky.h" />
     <ClInclude Include="Sphere.h" />
-    <ClInclude Include="Statue.h" />
+    <ClInclude Include="GameObjects\Statue.h" />
     <ClInclude Include="Texture.h" />
     <ClInclude Include="Textures\Textures.h" />
     <ClInclude Include="Timer.h" />
     <ClInclude Include="Vector.h" />
-    <ClInclude Include="Tunnel.h" />
+    <ClInclude Include="GameObjects\Tunnel.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
