+++ NonEuclidean/Scenes/Level1.h
@@ -1,7 +1,7 @@
 #pragma once
-#include "Scene.h"
+#include "../Scene.h"
 
-class Level5 : public Scene {
+class Level1 : public Scene {
 public:
   virtual void Load(PObjectVec& objs, PPortalVec& portals, Player& player) override;
 };
