+++ NonEuclidean/Scenes/Level2.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Scene.h"
+#include "../Scene.h"
 
 class Level2 : public Scene {
 public:
