+++ NonEuclidean/Scenes/Level4.cpp
@@ -1,6 +1,6 @@
 #include "Level4.h"
-#include "Tunnel.h"
-#include "Ground.h"
+#include "../GameObjects/Tunnel.h"
+#include "../GameObjects/Ground.h"
 
 void Level4::Load(PObjectVec& objs, PPortalVec& portals, Player& player) {
   std::shared_ptr<Tunnel> tunnel1(new Tunnel(Tunnel::SLOPE));
