+++ NonEuclidean/Scenes/Level4.h
@@ -1,7 +1,7 @@
 #pragma once
-#include "Scene.h"
+#include "../Scene.h"
 
-class Level3 : public Scene {
+class Level4 : public Scene {
 public:
   virtual void Load(PObjectVec& objs, PPortalVec& portals, Player& player) override;
 };
