+++ NonEuclidean/Scenes/Level6.cpp
@@ -1,5 +1,5 @@
 #include "Level6.h"
-#include "Floorplan.h"
+#include "../GameObjects/Floorplan.h"
 
 void Level6::Load(PObjectVec& objs, PPortalVec& portals, Player& player) {
   std::shared_ptr<Floorplan> floorplan(new Floorplan);
